feat: PDF ticket support
