+++ app/ai/ticket.py
@@ -1,8 +1,8 @@
 from loguru import logger
-from typing import Dict, Any
 
 from app.ai.prompts import ticket_info
 from app.ai.gemini import GeminiFileInformationExtractor
+from app.models import TicketInfo
 
 
 class TicketImageInformationExtractor(GeminiFileInformationExtractor):
@@ -10,12 +10,11 @@ class TicketImageInformationExtractor(GeminiFileInformationExtractor):
         super().__init__(api_key, model)
         self.prompt = ticket_info
 
-    def extract_ticket_info(self, file_data: bytes, mime_type: str) -> Dict[str, Any]:
+    async def extract_ticket_info(self, file_data: bytes, mime_type: str) -> TicketInfo:
         logger.info("Processing ticket from bytes")
         try:
-            file_uri = self.upload_file(file_data, mime_type)
-            extract_info = self.extract_info_from_file(file_uri, mime_type, self.prompt)
-            return extract_info
+            extract_info = await self.extract_ticket_info(file_data, mime_type)
+            return TicketInfo.parse_obj(extract_info)
         except Exception as e:
             logger.error(f"Error in extract_ticket_info: {str(e)}")
             raise
